fix(Editor) Preload assets where possible
